BigQuery/requirements.txt
kafka-python
pandas
google-cloud-bigquery